refactor: Phase 3.1g — extract Interaction, ViewPresets, StateAndExport mixins
Move 3 more method groups (529 lines) into separate mixin files:

InteractionMixin (main_window_interaction.py, 153 lines):
- _on_scroll, _on_mouse_press/release/motion
- _on_interaction_mode_changed, _reset_3d_view

ViewPresetsMixin (main_window_view_presets.py, 91 lines):
- _set_view_preset, _on_color_scheme_changed
- _get_color_scheme_colors, _get_transparency_values

StateAndExportMixin (main_window_state_export.py, 366 lines):
- get_current_state, restore_state
- set_electrode_depths, set_electrical_measurements
- export_results, _export_3d_plot, _export_charts

Monolith now at 1,016 lines (-76.8% from original 4,387).
10 mixins extracted total.
