+++ src/electrode_advisor/python/electrode_advisor/ui/pyqt6/main_window.py
@@ -15,7 +15,6 @@ from __future__ import annotations
 
 import logging
 import os
-import sys
 from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, cast
@@ -40,7 +39,6 @@ else:
 # Prevent matplotlib from opening separate windows
 mpl.rcParams["interactive"] = False  # Disable interactive mode
 
-import matplotlib.pyplot as plt  # noqa: E402
 import numpy as np  # noqa: E402
 
 if TYPE_CHECKING:
@@ -81,16 +79,8 @@ from upstream_drift_tools.calculators.electrical import (  # noqa: E402
     ThreePhaseElectricalModelEnhanced,
 )
 
-from ...configs.color_schemes import (  # noqa: E402
-    get_color_scheme,
-)
 from ...configs.ui_defaults import (  # noqa: E402
     PERIODIC_UPDATE_MS,
-)
-from ...configs.view_presets import (  # noqa: E402
-    DEFAULT_Z_SCALE_FACTOR,
-    DEFAULT_ZOOM_SCALE_FACTOR,
-    get_view_preset,
 )
 from ...ui_builders.visualization_builder import VisualizationBuilder  # noqa: E402
 from ...utils.visualization import ElectrodeVisualization  # noqa: E402
@@ -106,13 +96,19 @@ STATE_MIXIN_AVAILABLE = False
 from .main_window_drawing import DrawingMixin  # noqa: E402
 from .main_window_glass_integration import GlassIntegrationMixin  # noqa: E402
 from .main_window_input_panel import InputPanelMixin  # noqa: E402
+from .main_window_interaction import InteractionMixin  # noqa: E402
 from .main_window_results_charts import ResultsAndChartsMixin  # noqa: E402
+from .main_window_state_export import StateAndExportMixin  # noqa: E402
+from .main_window_view_presets import ViewPresetsMixin  # noqa: E402
 from .main_window_visual_controls import VisualControlsMixin  # noqa: E402
 from .main_window_visualization_update import VisualizationUpdateMixin  # noqa: E402
 
 
 # --- Main Widget for Tab Integration ---
 class ElectrodeAdvisorWidget(
+    StateAndExportMixin,
+    ViewPresetsMixin,
+    InteractionMixin,
     GlassIntegrationMixin,
     ResultsAndChartsMixin,
     VisualizationUpdateMixin,
@@ -1002,524 +998,9 @@ class ElectrodeAdvisorWidget(
         """Periodic update for real-time monitoring"""
         # This could refresh data from external sources
 
-    # Mouse interaction handlers
-    def _on_scroll(self, event: Any) -> None:
-        """Handle mouse scroll for zoom"""
-        try:
-            if event.inaxes == self.electrode_ax:
-                # Zoom factor
-                zoom_factor = 1.1 if event.button == "down" else 0.9
-
-                # Update zoom slider
-                current_zoom = self.zoom_slider.value()
-                new_zoom = int(current_zoom * zoom_factor)
-                new_zoom = max(50, min(200, new_zoom))  # Clamp to range
-                self.zoom_slider.setValue(new_zoom)
-
-        except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
-            logger.exception("Error in scroll handler: %s", e)
-
-    def _on_mouse_press(self, event: Any) -> None:
-        """Handle mouse press for rotation and pan"""
-        try:
-            if event.inaxes == self.electrode_ax:
-                self._mouse_pressed = True
-                # Use pixel coordinates for consistent behavior
-                self._last_mouse_pos = (
-                    (event.x, event.y)
-                    if event.x is not None and event.y is not None
-                    else None
-                )
-                # Get 3D view angles for rotation mode
-                if hasattr(self.electrode_ax, "elev") and hasattr(
-                    self.electrode_ax, "azim"
-                ):
-                    self._last_elev = self.electrode_ax.elev
-                    self._last_azim = self.electrode_ax.azim
-                else:
-                    # Fallback: use default angles if not available
-                    self._last_elev = 20
-                    self._last_azim = 45
-        except (ValueError, TypeError, ArithmeticError) as e:
-            logger.exception("Error in mouse press handler: %s", e)
-
-    def _on_mouse_release(self, event: Any) -> None:
-        """Handle mouse release"""
-        try:
-            self._mouse_pressed = False
-            logger.info("Mouse released")
-        except (ValueError, TypeError, ArithmeticError) as e:
-            logger.exception("Error in mouse release handler: %s", e)
-
-    def _on_mouse_motion(self, event: Any) -> None:
-        """Handle mouse motion for rotation or panning based on interaction mode"""
-        try:
-            if (
-                self._mouse_pressed
-                and event.inaxes == self.electrode_ax
-                and self._last_mouse_pos
-                and event.x is not None
-                and event.y is not None
-            ):
-                dx = event.x - self._last_mouse_pos[0]
-                dy = event.y - self._last_mouse_pos[1]
-                logger.info(
-                    "Mouse motion: dx=%s, dy=%s, mode=%s", dx, dy, self.interaction_mode
-                )
-                if self.interaction_mode == "rotation":
-                    # Only rotate if in rotation mode
-                    if hasattr(self.electrode_ax, "view_init") and hasattr(
-                        self, "_last_elev"
-                    ):
-                        rotation_sensitivity = 0.5
-                        new_elev = self._last_elev + dy * rotation_sensitivity
-                        new_azim = self._last_azim + dx * rotation_sensitivity
-                        new_elev = max(-90, min(90, new_elev))
-                        self.electrode_ax.view_init(elev=new_elev, azim=new_azim)
-                        if self.electrode_canvas:
-                            self.electrode_canvas.draw_idle()
-                        self._last_elev = new_elev
-                        self._last_azim = new_azim
-                elif self.interaction_mode == "pan":
-                    # Only pan if in pan mode
-                    if hasattr(self.electrode_ax, "get_xlim"):
-                        x_range = self.electrode_ax.get_xlim()
-                        y_range = self.electrode_ax.get_ylim()
-                        pan_factor = 0.01
-                        x_shift = dx * pan_factor * (x_range[1] - x_range[0])
-                        y_shift = dy * pan_factor * (y_range[1] - y_range[0])
-                        self.electrode_ax.set_xlim(
-                            x_range[0] - x_shift, x_range[1] - x_shift
-                        )
-                        self.electrode_ax.set_ylim(
-                            y_range[0] + y_shift, y_range[1] + y_shift
-                        )
-                    if self.electrode_canvas is not None:
-                        self.electrode_canvas.draw_idle()
-                self._last_mouse_pos = (event.x, event.y)
-        except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
-            logger.exception("Error in mouse motion handler: %s", e)
-
-    def _on_interaction_mode_changed(self) -> None:
-        """Handle interaction mode change"""
-        try:
-            if self.rotation_mode_radio.isChecked():
-                self.interaction_mode = "rotation"
-            else:
-                self.interaction_mode = "pan"
-            logger.info("3D interaction mode changed to: %s", self.interaction_mode)
-        except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
-            logger.exception("Error changing interaction mode: %s", e)
-
-    def _reset_3d_view(self) -> None:
-        """Reset 3D view to default angle and zoom"""
-        try:
-            if self.electrode_canvas is None or self.electrode_ax is None:
-                return
-
-            if hasattr(self.electrode_ax, "view_init"):
-                self.electrode_ax.view_init(elev=20, azim=45)
-
-            # Reset zoom slider to 100%
-            if hasattr(self, "zoom_slider"):
-                self.zoom_slider.setValue(100)
-
-            self.electrode_canvas.draw_idle()
-            logger.info("3D view reset to default")
-        except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
-            logger.exception("Error resetting 3D view: %s", e)
-
-    # Public methods for external integration
-    def _set_view_preset(self, preset: str) -> None:
-        """Set predefined view angles using view presets from config."""
-        if self.electrode_ax is None:
-            return
-
-        try:
-            view_angle = get_view_preset(preset)
-            self.electrode_ax.view_init(elev=view_angle.elev, azim=view_angle.azim)
-
-            # For default view, also reset pan/zoom
-            if preset == "default":
-                bath_diameter = self.bath_diameter_input.value()
-                extension_length = float(self.electrode_extension_slider.value())
-                glass_height = self.glass_layer_height_input.value()
-                metal_height = self.metal_layer_height_input.value()
-                max_range = max(
-                    bath_diameter / 2 + extension_length, glass_height + metal_height
-                )
-                zoom_factor = self.zoom_slider.value() / 100.0
-                scaled_range = max_range / zoom_factor * DEFAULT_ZOOM_SCALE_FACTOR
-                self.electrode_ax.set_xlim(-scaled_range, scaled_range)
-                self.electrode_ax.set_ylim(-scaled_range, scaled_range)
-                if hasattr(self.electrode_ax, "set_zlim"):
-                    self.electrode_ax.set_zlim(
-                        0,
-                        (glass_height + metal_height)
-                        / zoom_factor
-                        * DEFAULT_Z_SCALE_FACTOR,
-                    )
-
-            if self.electrode_canvas is not None:
-                self.electrode_canvas.draw()
-        except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
-            logger.exception("Error setting view preset: %s", e)
-
-    def _on_color_scheme_changed(self, scheme: str) -> None:
-        """Handle color scheme changes"""
-        try:
-            self.current_color_scheme = scheme
-            # Trigger visualization update with new color scheme
-            self._on_input_changed()
-        except (ValueError, TypeError, ArithmeticError) as e:
-            logger.exception("Error changing color scheme: %s", e)
-
-    def _get_color_scheme_colors(self) -> list[str]:
-        """Get colors based on current color scheme."""
-        return get_color_scheme(self.current_color_scheme)
-
-    def _get_transparency_values(self) -> dict[str, float]:
-        """Get current transparency values from sliders"""
-        return {
-            "electrodes": self.electrode_alpha_slider.value() / 100.0,
-            "glass": self.glass_alpha_slider.value() / 100.0,
-            "metal": self.metal_alpha_slider.value() / 100.0,
-            "paths": self.path_alpha_slider.value() / 100.0,
-            "refractory": self.refractory_alpha_slider.value() / 100.0,
-            "metal_shell": self.metal_shell_alpha_slider.value() / 100.0,
-        }
-
-    def get_current_state(self) -> dict[str, Any]:
-        """Get current state for state management"""
-        try:
-            return {
-                "calculator_name": "ElectrodeAdvisor",
-                "timestamp": datetime.now().isoformat(),
-                "config": {
-                    "bath_radius": self.bath_diameter_input.value() / 2.0,
-                    "electrode_radius": float(
-                        self.electrode_diameter_combo.currentText()
-                    )
-                    / 2.0,
-                    "metal_height": self.config.metal_layer_height,
-                    "glass_height": self.config.glass_depth,
-                    "bath_temperature": self.config.bath_temperature_base,
-                    "vertical_spreading": self.config.vertical_spreading_factor,
-                    "horizontal_spreading": self.config.horizontal_spreading_factor,
-                },
-                "electrode_depths": {
-                    "E1": (
-                        getattr(self, "depth_inputs", [None, None, None])[0].value()
-                        if hasattr(self, "depth_inputs") and len(self.depth_inputs) > 0
-                        else 0
-                    ),
-                    "E2": (
-                        getattr(self, "depth_inputs", [None, None, None])[1].value()
-                        if hasattr(self, "depth_inputs") and len(self.depth_inputs) > 1
-                        else 0
-                    ),
-                    "E3": (
-                        getattr(self, "depth_inputs", [None, None, None])[2].value()
-                        if hasattr(self, "depth_inputs") and len(self.depth_inputs) > 2
-                        else 0
-                    ),
-                },
-                "electrical_measurements": {
-                    "1-2": {
-                        "current": (
-                            getattr(self, "phase_inputs", {})
-                            .get("1-2", {})
-                            .get("current", None)
-                            .value()
-                            if hasattr(self, "phase_inputs")
-                            and "1-2" in self.phase_inputs
-                            else 0
-                        ),
-                        "voltage": (
-                            getattr(self, "phase_inputs", {})
-                            .get("1-2", {})
-                            .get("voltage", None)
-                            .value()
-                            if hasattr(self, "phase_inputs")
-                            and "1-2" in self.phase_inputs
-                            else 0
-                        ),
-                    },
-                    "2-3": {
-                        "current": (
-                            getattr(self, "phase_inputs", {})
-                            .get("2-3", {})
-                            .get("current", None)
-                            .value()
-                            if hasattr(self, "phase_inputs")
-                            and "2-3" in self.phase_inputs
-                            else 0
-                        ),
-                        "voltage": (
-                            getattr(self, "phase_inputs", {})
-                            .get("2-3", {})
-                            .get("voltage", None)
-                            .value()
-                            if hasattr(self, "phase_inputs")
-                            and "2-3" in self.phase_inputs
-                            else 0
-                        ),
-                    },
-                    "3-1": {
-                        "current": (
-                            getattr(self, "phase_inputs", {})
-                            .get("3-1", {})
-                            .get("current", None)
-                            .value()
-                            if hasattr(self, "phase_inputs")
-                            and "3-1" in self.phase_inputs
-                            else 0
-                        ),
-                        "voltage": (
-                            getattr(self, "phase_inputs", {})
-                            .get("3-1", {})
-                            .get("voltage", None)
-                            .value()
-                            if hasattr(self, "phase_inputs")
-                            and "3-1" in self.phase_inputs
-                            else 0
-                        ),
-                    },
-                },
-                "results": self.calculation_results,
-            }
-        except (ValueError, ZeroDivisionError, OverflowError, TypeError) as e:
-            logger.exception("Error getting current state: %s", e)
-            return {"calculator_name": "ElectrodeAdvisor", "error": str(e)}
-
-    def restore_state(self, state_data: dict[str, Any]) -> bool:
-        """Restore state from saved data"""
-        try:
-            if state_data.get("calculator_name") != "ElectrodeAdvisor":
-                logger.info("State data mismatch for ElectrodeAdvisor")
-                return False
-
-            # Restore configuration
-            if "config" in state_data:
-                config_data = state_data["config"]
-                if "bath_radius" in config_data:
-                    self.bath_diameter_input.setValue(config_data["bath_radius"] * 2.0)
-                if "electrode_radius" in config_data:
-                    self.electrode_diameter_combo.setCurrentText(
-                        str(config_data["electrode_radius"] * 2.0)
-                    )
-                if "metal_height" in config_data:
-                    self.config.metal_layer_height = config_data["metal_height"]
-                if "glass_height" in config_data:
-                    self.config.glass_depth = config_data["glass_height"]
-                if "bath_temperature" in config_data:
-                    self.config.bath_temperature_base = config_data["bath_temperature"]
-                if "vertical_spreading" in config_data:
-                    self.config.vertical_spreading_factor = config_data[
-                        "vertical_spreading"
-                    ]
-                if "horizontal_spreading" in config_data:
-                    self.config.horizontal_spreading_factor = config_data[
-                        "horizontal_spreading"
-                    ]
-
-            # Restore electrode depths
-            if "electrode_depths" in state_data and hasattr(self, "depth_inputs"):
-                depths_data = state_data["electrode_depths"]
-                if "E1" in depths_data and len(self.depth_inputs) > 0:
-                    self.depth_inputs[0].setValue(depths_data["E1"])
-                if "E2" in depths_data and len(self.depth_inputs) > 1:
-                    self.depth_inputs[1].setValue(depths_data["E2"])
-                if "E3" in depths_data and len(self.depth_inputs) > 2:
-                    self.depth_inputs[2].setValue(depths_data["E3"])
-
-            # Restore electrical measurements
-            if "electrical_measurements" in state_data and hasattr(
-                self, "phase_inputs"
-            ):
-                measurements_data = state_data["electrical_measurements"]
-                for phase in ["1-2", "2-3", "3-1"]:
-                    if phase in measurements_data and phase in self.phase_inputs:
-                        if "current" in measurements_data[phase]:
-                            cast(
-                                QDoubleSpinBox, self.phase_inputs[phase]["current"]
-                            ).setValue(measurements_data[phase]["current"])
-                        if "voltage" in measurements_data[phase]:
-                            cast(
-                                QDoubleSpinBox, self.phase_inputs[phase]["voltage"]
-                            ).setValue(measurements_data[phase]["voltage"])
-
-            # Trigger recalculation
-            QTimer.singleShot(100, self._calculate_system)
-
-            logger.info("State restored for ElectrodeAdvisor")
-            return True
-
-        except (RuntimeError, AttributeError) as e:
-            logger.exception("Error restoring state for ElectrodeAdvisor: %s", e)
-            return False
-
-    def set_electrode_depths(self, depths: list[float]) -> None:
-        """Set electrode depths programmatically"""
-        for i, depth in enumerate(depths[:3]):
-            self.depth_inputs[i].setValue(depth)
-
-    def set_electrical_measurements(
-        self, currents: list[float], voltages: list[float]
-    ) -> None:
-        """Set electrical measurements programmatically"""
-        phases = ["1-2", "2-3", "3-1"]
-        for i, phase in enumerate(phases[:3]):
-            if phase in self.phase_inputs:
-                cast(QDoubleSpinBox, self.phase_inputs[phase]["current"]).setValue(
-                    currents[i]
-                )
-                cast(QDoubleSpinBox, self.phase_inputs[phase]["voltage"]).setValue(
-                    voltages[i]
-                )
-
-    def export_results(self, filename: str) -> bool:
-        """Export calculation results to file"""
-        try:
-            import json
-
-            with open(filename, "w") as f:
-                # Prepare exportable data
-                export_data = {
-                    "timestamp": datetime.now().isoformat(),
-                    "configuration": {
-                        "bath_diameter": self.bath_diameter_input.value(),
-                        "electrode_diameter": float(
-                            self.electrode_diameter_combo.currentText()
-                        ),
-                        "metal_height": self.metal_layer_height_input.value(),
-                        "glass_height": self.glass_layer_height_input.value(),
-                        "bath_temperature": self.bath_temp_input.value(),
-                        "vertical_spreading": self.config.vertical_spreading_factor,
-                        "horizontal_spreading": self.config.horizontal_spreading_factor,
-                    },
-                    "electrode_depths": {
-                        "E1": self.depth_inputs[0].value(),
-                        "E2": self.depth_inputs[1].value(),
-                        "E3": self.depth_inputs[2].value(),
-                    },
-                    "results": self.calculation_results,
-                }
-                json.dump(export_data, f, indent=2)
-                logger.info("Results exported to %s", filename)
-                return True
-        except (PermissionError, OSError) as e:
-            logger.exception("Error exporting results: %s", e)
-            return False
-
-    def _export_3d_plot(self) -> None:
-        """Export the 3D plot to an image file"""
-        try:
-            from PyQt6.QtWidgets import QFileDialog
-
-            filename, _ = QFileDialog.getSaveFileName(
-                self,
-                "Export 3D Plot",
-                "electrode_3d_plot.png",
-                "PNG files (*.png);;PDF files (*.pdf);;SVG files (*.svg);;JPG files (*.jpg)",
-            )
-
-            if filename and self.electrode_fig is not None:
-                # Save the main 3D plot
-                self.electrode_fig.savefig(filename, dpi=300, bbox_inches="tight")
-                self._update_status(f"3D plot exported to {filename}", "ok")
-
-        except ImportError as e:
-            self._update_status(f"Error exporting 3D plot: {e!s}", "error")
-            logger.exception("Export error: %s", e)
-
-    def _export_charts(self) -> None:
-        """Export the current and power charts to image files"""
-        try:
-            from PyQt6.QtWidgets import QFileDialog
-
-            filename, _ = QFileDialog.getSaveFileName(
-                self,
-                "Export Charts",
-                "electrode_charts.png",
-                "PNG files (*.png);;PDF files (*.pdf);;SVG files (*.svg);;JPG files (*.jpg)",
-            )
-
-            if filename:
-                # Create a combined figure with both charts
-                # Performance: Use Figure() directly instead of plt.figure() to avoid
-                # pyplot state machine memory leaks in long-running sessions
-                combined_fig = Figure(
-                    figsize=(6, 3)
-                )  # REDUCED: Smaller figure size to allow flexible sizing
-
-                # Copy current chart
-                ax1 = combined_fig.add_subplot(1, 2, 1)
-
-                # Recreate current distribution chart
-                phase_currents = [
-                    self.phase_inputs["1-2"]["current"].value(),
-                    self.phase_inputs["2-3"]["current"].value(),
-                    self.phase_inputs["3-1"]["current"].value(),
-                ]
-                phases = ["1-2", "2-3", "3-1"]
-                colors = ["#FF4444", "#44FF44", "#4444FF"]
-                x = np.arange(len(phases))
-                width = 0.35
-
-                ax1.bar(
-                    x - width / 2,
-                    phase_currents,
-                    width,
-                    label="Phase Current",
-                    color=colors,
-                    alpha=0.8,
-                )
-                line_currents = [current * 0.8 for current in phase_currents]
-                ax1.bar(
-                    x + width / 2,
-                    line_currents,
-                    width,
-                    label="Line Current",
-                    color=colors,
-                    alpha=0.5,
-                )
-
-                ax1.set_xlabel("Phase")
-                ax1.set_ylabel("Current (A)")
-                ax1.set_title("Current Distribution")
-                ax1.set_xticks(x)
-                ax1.set_xticklabels(phases)
-                ax1.legend(loc="upper right")
-                ax1.grid(True, alpha=0.3)
-
-                # Copy power chart
-                ax2 = combined_fig.add_subplot(1, 2, 2)
-
-                # Recreate power distribution chart
-                powers = []
-                colors2 = ["#FF8C00", "#32CD32", "#1E90FF"]
-                for phase in phases:
-                    current = self.phase_inputs[phase]["current"].value()
-                    voltage = self.phase_inputs[phase]["voltage"].value()
-                    power = current * voltage / 1000
-                    powers.append(power)
-
-                ax2.bar(phases, powers, color=colors2, alpha=0.7)
-                ax2.set_xlabel("Phase")
-                ax2.set_ylabel("Power (kW)")
-                ax2.set_title("Power Distribution")
-                ax2.grid(True, alpha=0.3)
-
-                combined_fig.tight_layout()
-                combined_fig.savefig(filename, dpi=300, bbox_inches="tight")
-                plt.close(combined_fig)
-
-                self._update_status(f"Charts exported to {filename}", "ok")
-
-        except ImportError as e:
-            self._update_status(f"Error exporting charts: {e!s}", "error")
-            logger.exception("Export error: %s", e)
+    # Mouse/interaction methods -> InteractionMixin (main_window_interaction.py)
+    # View presets/color scheme -> ViewPresetsMixin (main_window_view_presets.py)
+    # State/export methods -> StateAndExportMixin (main_window_state_export.py)
 
 
 # For backward compatibility with the wrapper
